requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
colorama==0.4.6
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
pyaes==1.6.1
pymediainfo-pyrofork==6.0.2
pyrofork==2.3.68
PySocks==1.7.1
python-dotenv==1.1.1
TgCrypto==1.2.5
yarl==1.20.1
